Fix 697461: Use floating point resolution in mupdf-x11.
Fitting the page to the screen needs more precision than an integer DPI.

+++ platform/x11/pdfapp.h
@@ -70,7 +70,7 @@ struct pdfapp_s
 	int pagecount;
 
 	/* current view params */
-	int resolution;
+	float resolution;
 	int rotate;
 	fz_pixmap *image;
 	int grayscale;
